third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/iframe-reload-expected.txt

This is a testharness.js-based test.
PASS Reloaded same-host fetches are strictly same-site
PASS Reloaded subdomain fetches are strictly same-site
FAIL Reloaded cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.18461347777392034" but got (undefined) undefined
Harness: the test ran to completion.
